requirements_doc_level_search.txt
chromadb==0.4.21
sentence-transformers==2.3.1
pymongo==4.6.1
langchain>=0.1.0
langchain-core>=0.1.20
langchain-anthropic>=0.1.4
python-dotenv>=1.0.0 